feat(packages/web): separate awesome routes
